arg_heuristic(): be better prepared to handle path lists
If we were to handle only single paths, the buffer really would only
need to fit a single one, but arg_heuristic is also used to convert e.g.
INFO_PATH and MAN_PATH (which can contain multiple absolute POSIX-style
paths that need to be converted to Windows-style paths).

This is only really relevant if the POSIX root directory's path is
already pretty long. Such a beast has been spotted in the wild, in the
Git for Windows project:

	https://github.com/git-for-windows/git/issues/303

Of course, this patch is only a work-around that helps in many, but not
all cases. Obviously this patch will wait patiently until a real fix
comes around the corner, and then readily retire.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>